delete old component & add columns file for visas view
