--- src/components/specific/files/files-table/visas-view/VisasView.scss
@@ -1,77 +1,0 @@
-.visas-view {
-  &__element,
-  &__header {
-    & > * {
-      padding: 0 var(--spacing-unit);
-      margin-right: 2.5px;
-      height: 100%;
-
-      flex-shrink: 0;
-      text-align: center;
-      align-content: center;
-    }
-  }
-  .validation-tag {
-    margin: auto;
-    width: max-content;
-    height: 25px;
-    padding: 0 var(--spacing-unit);
-    border-radius: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 12px;
-    &.closed-visa {
-      color: var(--color-granite-light);
-      background-color: #ebebeb;
-    }
-    &.ask-visa {
-      color: #205dbd;
-      background-color: #e9eff8;
-    }
-    &.waiting-visa {
-      color: #ff9100;
-      background-color: #fef9ea;
-    }
-  }
-  &__validator {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    img {
-      width: 28px;
-      height: 28px;
-      border-radius: 50px;
-    }
-  }
-  &__status {
-    margin: auto;
-    padding: 0 var(--spacing-unit);
-    height: 26px;
-    width: max-content;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 50px;
-    &.success {
-      background-color: #e5f7ed;
-      color: var(--color-success);
-    }
-    &.failed {
-      background-color: var(--color-high-lighter);
-      color: var(--color-high);
-    }
-    &.sandglass {
-      background-color: rgba(47, 55, 75, 0.05);
-      color: var(--color-primary);
-    }
-    &.delay {
-      background-color: #fff4e5;
-      color: var(--color-warning);
-    }
-  }
-  .validate-visa-btn {
-    color: #205dbd;
-    border-color: #205dbd;
-  }
-}
